Run Gibbs sampler for covariates to use in network causal effect estimation

Given the specific inputs, determine covariate values using a Gibbs sampling procedure.

networkGibbsOutCovCpp(tau, rho, nu, ncov, R, N, burnin, rho_mat, adjacency,
  weights, cov_mat, group_lengths, group_functions)

Arguments

tau
A numeric vector for the intercept terms in the covariate model

rho
A numeric vector for the correlation terms in the covariate model

nu
A numberic vector for the neighbor terms in the covariate model

ncov
An integer for the number of covariates

R
An integer indicating the number of iterations for the Gibbs

N
An integer indicating the size of the interconnected network

burnin
An integer indicating when to start saving values in the chain

rho_mat
A numeric matrix for rho terms

adjacency
A binary matrix indicating connected units

weights
A numeric vector indicating the number of neighbors for each node

cov_mat
A numeric matrix for starting values for each covariate

group_lengths
An integer vector indicating the number of categories for each variable

group_functions
An integer vector indicating the type of variable

Value

A list of numeric matrices that contain the covariate values and neighbor covariate values for each person at that specific point in the chain